+++ src/index.ts
@@ -100,6 +100,7 @@ export * from './components/progress-circular';
 export * from './components/progress-linear';
 export * from './components/radio-group';
 export * from './components/rating';
+export * from './components/scatter-chart';
 export * from './components/scroll-area';
 export * from './components/scroll-zone';
 export * from './components/segmented-button';